Rename zustand-stores folder to stores
